AIEngineering/ai-pr-learning/requirements.txt
requests>=2.32.0
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
